@@ -2,7 +2,7 @@
 name = "catlog-wasm"
 version = "0.1.0"
 authors = ["Evan Patterson <evan@epatters.org>"]
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
@@ -14,14 +14,18 @@ default = ["console_error_panic_hook"]
 all-the-same = "1.1.0"
 catlog = { path = "../catlog", features = ["ode", "serde-wasm"] }
 console_error_panic_hook = { version = "0.1.7", optional = true }
-derive_more = { version = "1", features = ["from", "try_into"] }
+derive_more = { version = "2", features = ["from", "try_into"] }
+egglog = { version = "0.4", default-features = false, features = [
+    "wasm-bindgen",
+] }
 getrandom = { version = "0.2", features = ["js"] }
 js-sys = "0.3.77"
 nonempty = { version = "0.10", features = ["serialize"] }
+notebook-types = { version = "0.1.0", path = "../notebook-types" }
 serde = { version = "1", features = ["derive"] }
-tsify-next = { version = "0.5.4", features = ["js"] }
+tsify = { version = "0.5", features = ["js"] }
 ustr = "1"
-uuid = { version = "1.10", features = ["v7", "serde", "js"] }
+uuid = { version = "=1.11", features = ["v7", "serde", "js"] }
 wasm-bindgen = "0.2.100"
 
 [dev-dependencies]
