@@ -3,20 +3,20 @@ name = "catlog"
 description = "A toolbox for catgorical logic based on double theories"
 authors = ["Evan Patterson"]
 version = "0.1.0"
-edition = "2021"
-rust-version = "1.76"
+edition = "2024"
 
 [features]
 ode = ["dep:ode_solvers", "dep:nalgebra"]
 serde = ["dep:serde", "nonempty/serialize", "ustr/serde"]
-serde-wasm = ["serde", "dep:wasm-bindgen", "dep:tsify-next"]
+serde-wasm = ["serde", "dep:wasm-bindgen", "dep:tsify"]
 
 [dependencies]
-archery = "1"
 derivative = "2"
-derive_more = { version = "1", features = ["from", "into"] }
+derive_more = { version = "2", features = ["constructor", "from", "into"] }
 duplicate = "2"
-either = "1"
+egglog = { version = "0.4", default-features = false }
+ego-tree = "0.10"
+itertools = "0.14"
 nalgebra = { version = "0.33", optional = true }
 nonempty = "0.10"
 num-traits = "0.2"
@@ -24,8 +24,9 @@ ode_solvers = { version = "0.5.0", optional = true }
 ref-cast = "1"
 serde = { version = "1", features = ["derive"], optional = true }
 thiserror = "1"
-tsify-next = { version = "0.5", features = ["js"], optional = true }
+tsify = { version = "0.5", features = ["js"], optional = true }
 ustr = "1"
+uuid = { version = "=1.11", optional = true }
 wasm-bindgen = { version = "0.2.100", optional = true }
 
 [dev-dependencies]
