@@ -50,6 +50,35 @@ impl Permissions {
     }
 }
 
+/// Returns an error if the user_id in the session does not exist in the DB, returns None otherwise
+///
+/// Used by the client to gracefully handle stale sessions
+pub async fn validate_session(ctx: AppCtx) -> Result<(), AppError> {
+    let user_id = match ctx.user.as_ref().map(|u| u.user_id.clone()) {
+        Some(id) => id,
+        None => {
+            return Ok(());
+        }
+    };
+
+    let exists = sqlx::query_scalar!(
+        r#"
+        SELECT EXISTS (
+            SELECT 1 FROM users WHERE id = $1
+        )
+        "#,
+        user_id
+    )
+    .fetch_one(&ctx.state.db)
+    .await?;
+
+    if !exists.unwrap_or(false) {
+        return Err(AppError::Unauthorized);
+    }
+
+    Ok(())
+}
+
 /** Verify that user is authorized to access a ref at a given permission level.
 
 It is safe to proceed if the result is `Ok`; otherwise, the requested action
@@ -252,7 +281,7 @@ pub fn authenticate_from_request<T>(
 
             firebase_auth
                 .verify(bearer)
-                .map_err(|err| format!("Failed to verify token: {}", err))
+                .map_err(|err| format!("Failed to verify token: {err}"))
         })
         .transpose()
 }
