@@ -1,29 +1,38 @@
+import { runMigration } from "./migrations/index.js";
 import { AutomergeServer } from "./server.js";
 import { SocketServer } from "./socket.js";
 
-const internal_port = process.env.AUTOMERGE_INTERNAL_PORT || 3000;
-const port = process.env.AUTOMERGE_PORT || 8010;
+async function main() {
+    const args = process.argv.slice(2);
 
-const server = new AutomergeServer(port);
+    if (args[0] === "--migrate") {
+        const migrationName = args[1];
+        if (!migrationName) {
+            console.error("Missing migration name after --migrate");
+            process.exit(1);
+        }
 
-const socket_server = new SocketServer(internal_port, {
-    createDoc(data) {
-        const { refId, content } = data;
-        return server.createDocHandle(refId, content).documentId;
-    },
-    getDoc(refId) {
-        return server.getDocHandle(refId)?.documentId ?? null;
-    },
-});
+        await runMigration(migrationName);
+        process.exit(0);
+    }
 
-server.handleChange = (refId, content) => socket_server.autosave(refId, content);
+    const internal_port = process.env.AUTOMERGE_INTERNAL_PORT || 3000;
+    const port = process.env.AUTOMERGE_PORT || 8010;
 
-process.once("SIGINT", () => {
-    socket_server.close();
-    server.close();
-});
+    const server = new AutomergeServer(port);
+    const socket_server = new SocketServer(internal_port, server);
 
-process.once("SIGTERM", () => {
-    socket_server.close();
-    server.close();
-});
+    server.handleChange = (refId, content) => socket_server.autosave(refId, content);
+
+    process.once("SIGINT", () => {
+        socket_server.close();
+        server.close();
+    });
+
+    process.once("SIGTERM", () => {
+        socket_server.close();
+        server.close();
+    });
+}
+
+main();
