@@ -4,27 +4,27 @@ Each struct in this module provides a [`DblTheory`] plus possibly
 theory-specific analysis methods.
  */
 
-use std::sync::Arc;
+use std::rc::Rc;
 
 use ustr::ustr;
 use wasm_bindgen::prelude::*;
 
-use catlog::dbl::{model::DiscreteDblModel, theory};
-use catlog::one::fin_category::FinMor;
+use catlog::dbl::{model, theory};
+use catlog::one::Path;
 use catlog::stdlib::{analyses, models, theories};
 
-use super::model_morphism::motifs;
+use super::model_morphism::{MotifsOptions, motifs};
 use super::{analyses::*, model::DblModel, theory::DblTheory};
 
 /// The empty or initial theory.
 #[wasm_bindgen]
-pub struct ThEmpty(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThEmpty(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThEmpty {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_empty()))
+        Self(Rc::new(theories::th_empty()))
     }
 
     #[wasm_bindgen]
@@ -35,13 +35,13 @@ impl ThEmpty {
 
 /// The theory of categories.
 #[wasm_bindgen]
-pub struct ThCategory(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThCategory(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThCategory {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_category()))
+        Self(Rc::new(theories::th_category()))
     }
 
     #[wasm_bindgen]
@@ -52,13 +52,13 @@ impl ThCategory {
 
 /// The theory of database schemas with attributes.
 #[wasm_bindgen]
-pub struct ThSchema(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThSchema(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThSchema {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_schema()))
+        Self(Rc::new(theories::th_schema()))
     }
 
     #[wasm_bindgen]
@@ -69,13 +69,13 @@ impl ThSchema {
 
 /// The theory of signed categories.
 #[wasm_bindgen]
-pub struct ThSignedCategory(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThSignedCategory(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThSignedCategory {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_signed_category()))
+        Self(Rc::new(theories::th_signed_category()))
     }
 
     #[wasm_bindgen]
@@ -85,34 +85,64 @@ impl ThSignedCategory {
 
     /// Find positive feedback loops in a model.
     #[wasm_bindgen(js_name = "positiveLoops")]
-    pub fn positive_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn positive_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let positive_loop = models::positive_loop(self.0.clone());
-        motifs(&positive_loop, model)
+        motifs(&positive_loop, model, options)
     }
 
     /// Find negative feedback loops in a model.
     #[wasm_bindgen(js_name = "negativeLoops")]
-    pub fn negative_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn negative_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let negative_loop = models::negative_loop(self.0.clone());
-        motifs(&negative_loop, model)
+        motifs(&negative_loop, model, options)
     }
 
-    /// Simulate Lotka-Volterra system derived from a model.
+    /// Simulate the Lotka-Volterra system derived from a model.
     #[wasm_bindgen(js_name = "lotkaVolterra")]
     pub fn lotka_volterra(
         &self,
         model: &DblModel,
         data: LotkaVolterraModelData,
     ) -> Result<ODEResult, String> {
-        let model: &DiscreteDblModel<_, _> = (&model.0)
+        let model: &model::DiscreteDblModel<_, _> = (&model.0)
             .try_into()
             .map_err(|_| "Lotka-Volterra simulation expects a discrete double model")?;
         Ok(ODEResult(
             analyses::ode::LotkaVolterraAnalysis::new(ustr("Object"))
-                .add_positive(FinMor::Id(ustr("Object")))
-                .add_negative(FinMor::Generator(ustr("Negative")))
-                .solve(model, data.0)
-                .map_err(|err| format!("{:?}", err))
+                .add_positive(Path::Id(ustr("Object")))
+                .add_negative(ustr("Negative").into())
+                .create_system(model, data.0)
+                .solve_with_defaults()
+                .map_err(|err| format!("{err:?}"))
+                .into(),
+        ))
+    }
+
+    /// Simulate the linear ODE system derived from a model.
+    #[wasm_bindgen(js_name = "linearODE")]
+    pub fn linear_ode(
+        &self,
+        model: &DblModel,
+        data: LinearODEModelData,
+    ) -> Result<ODEResult, String> {
+        let model: &model::DiscreteDblModel<_, _> = (&model.0)
+            .try_into()
+            .map_err(|_| "Linear ODE simulation expects a discrete double model")?;
+        Ok(ODEResult(
+            analyses::ode::LinearODEAnalysis::new(ustr("Object"))
+                .add_positive(Path::Id(ustr("Object")))
+                .add_negative(ustr("Negative").into())
+                .create_system(model, data.0)
+                .solve_with_defaults()
+                .map_err(|err| format!("{err:?}"))
                 .into(),
         ))
     }
@@ -120,13 +150,13 @@ impl ThSignedCategory {
 
 /// The theory of delayable signed categories.
 #[wasm_bindgen]
-pub struct ThDelayableSignedCategory(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThDelayableSignedCategory(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThDelayableSignedCategory {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_delayable_signed_category()))
+        Self(Rc::new(theories::th_delayable_signed_category()))
     }
 
     #[wasm_bindgen]
@@ -136,42 +166,58 @@ impl ThDelayableSignedCategory {
 
     /// Find (fast) positive feedback loops in a model.
     #[wasm_bindgen(js_name = "positiveLoops")]
-    pub fn positive_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn positive_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let positive_loop = models::positive_loop(self.0.clone());
-        motifs(&positive_loop, model)
+        motifs(&positive_loop, model, options)
     }
 
     /// Find (fast) negative feedback loops in a model.
     #[wasm_bindgen(js_name = "negativeLoops")]
-    pub fn negative_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn negative_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let negative_loop = models::negative_loop(self.0.clone());
-        motifs(&negative_loop, model)
+        motifs(&negative_loop, model, options)
     }
 
     /// Find delayed positive feedback loops in a model.
     #[wasm_bindgen(js_name = "delayedPositiveLoops")]
-    pub fn delayed_positive_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn delayed_positive_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let delayed_positive_loop = models::delayed_positive_loop(self.0.clone());
-        motifs(&delayed_positive_loop, model)
+        motifs(&delayed_positive_loop, model, options)
     }
 
     /// Find delayed negative feedback loops in a model.
     #[wasm_bindgen(js_name = "delayedNegativeLoops")]
-    pub fn delayed_negative_loops(&self, model: &DblModel) -> Result<Vec<DblModel>, String> {
+    pub fn delayed_negative_loops(
+        &self,
+        model: &DblModel,
+        options: MotifsOptions,
+    ) -> Result<Vec<DblModel>, String> {
         let delayed_negative_loop = models::delayed_negative_loop(self.0.clone());
-        motifs(&delayed_negative_loop, model)
+        motifs(&delayed_negative_loop, model, options)
     }
 }
 
 /// The theory of nullable signed categories.
 #[wasm_bindgen]
-pub struct ThNullableSignedCategory(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThNullableSignedCategory(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThNullableSignedCategory {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_nullable_signed_category()))
+        Self(Rc::new(theories::th_nullable_signed_category()))
     }
 
     #[wasm_bindgen]
@@ -182,13 +228,13 @@ impl ThNullableSignedCategory {
 
 /// The theory of categories with scalars.
 #[wasm_bindgen]
-pub struct ThCategoryWithScalars(Arc<theory::UstrDiscreteDblTheory>);
+pub struct ThCategoryWithScalars(Rc<theory::UstrDiscreteDblTheory>);
 
 #[wasm_bindgen]
 impl ThCategoryWithScalars {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_category_with_scalars()))
+        Self(Rc::new(theories::th_category_with_scalars()))
     }
 
     #[wasm_bindgen]
@@ -199,18 +245,37 @@ impl ThCategoryWithScalars {
 
 /// The theory of categories with links.
 #[wasm_bindgen]
-pub struct ThCategoryLinks(Arc<theory::UstrDiscreteTabTheory>);
+pub struct ThCategoryLinks(Rc<theory::UstrDiscreteTabTheory>);
 
 #[wasm_bindgen]
 impl ThCategoryLinks {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
-        Self(Arc::new(theories::th_category_links()))
+        Self(Rc::new(theories::th_category_links()))
     }
 
     #[wasm_bindgen]
     pub fn theory(&self) -> DblTheory {
         DblTheory(self.0.clone().into())
+    }
+
+    /// Simulates the mass-action system derived from a model.
+    #[wasm_bindgen(js_name = "massAction")]
+    pub fn mass_action(
+        &self,
+        model: &DblModel,
+        data: MassActionModelData,
+    ) -> Result<ODEResult, String> {
+        let model: &model::DiscreteTabModel<_, _, _> = (&model.0)
+            .try_into()
+            .map_err(|_| "Mass-action simulation expects a discrete tabulator model")?;
+        Ok(ODEResult(
+            analyses::ode::StockFlowMassActionAnalysis::default()
+                .create_numerical_system(model, data.0)
+                .solve_with_defaults()
+                .map_err(|err| format!("{err:?}"))
+                .into(),
+        ))
     }
 }
 
