@@ -1,11 +1,17 @@
-//! Category theory in dimension one.
+//! Category theory in dimension one, plus a little graph theory.
 
 pub mod category;
-pub mod fin_category;
+pub mod computad;
+pub mod fp_category;
+pub mod functor;
 pub mod graph;
 pub mod graph_algorithms;
 pub mod path;
+pub mod tree;
+pub mod tree_algorithms;
 
 pub use self::category::*;
+pub use self::fp_category::*;
+pub use self::functor::*;
 pub use self::graph::*;
 pub use self::path::*;
