@@ -1,11 +1,9 @@
-//! Sets and functions, and a little abstract algebra.
+//! Sets and functions, plus a little abstract algebra.
 
 pub mod alg;
 pub mod column;
-pub mod directories;
 pub mod rig;
 pub mod set;
 
 pub use self::column::*;
-pub use self::directories::*;
 pub use self::set::*;
