@@ -1,3 +1,3 @@
-export * from "./types";
 export * from "./document";
 export * from "./context";
+export * from "./types";
