@@ -19,6 +19,7 @@ import {
     onCleanup,
 } from "solid-js";
 
+import type { Cell, Notebook } from "catlog-wasm";
 import { type Completion, IconButton } from "../components";
 import { deepCopyJSON } from "../util/deepcopy";
 import {
@@ -29,15 +30,8 @@ import {
     StemCellEditor,
     isCellDragData,
 } from "./notebook_cell";
-import {
-    type Cell,
-    type FormalCell,
-    type Notebook,
-    newFormalCell,
-    newRichTextCell,
-    newStemCell,
-} from "./types";
 import { WalkthroughOverlay } from "./walkthrough_overlay";
+import { type FormalCell, newFormalCell, newRichTextCell, newStemCell } from "./types";
 
 import "./notebook_editor.css";
 
