@@ -1,23 +1,6 @@
 import { v7 } from "uuid";
 
-import type { Uuid } from "catlog-wasm";
-
-/** Data type for a notebook.
-
-A notebook is nothing more than a list of cells. Any metadata associated with
-notebook, such as its title, is stored elsewhere.
- */
-export type Notebook<T> = {
-    cells: Cell<T>[];
-};
-
-/** A cell in a notebook.
-
-Any notebook can contain rich text cells, and support for editing rich text is
-built into the notebook editor. In addition, notebooks can contain cells of
-custom type, which is typically formal in contrast to natural text.
- */
-export type Cell<T> = RichTextCell | FormalCell<T> | StemCell;
+import type { Cell, Notebook } from "catlog-wasm";
 
 /** Creates an empty notebook. */
 export const newNotebook = <T>(): Notebook<T> => ({
@@ -25,11 +8,7 @@ export const newNotebook = <T>(): Notebook<T> => ({
 });
 
 /** A cell containing rich text. */
-export type RichTextCell = {
-    tag: "rich-text";
-    id: Uuid;
-    content: string;
-};
+export type RichTextCell = Cell<unknown> & { tag: "rich-text" };
 
 /** Creates a rich text cell with the given content. */
 export const newRichTextCell = (content?: string): RichTextCell => ({
@@ -39,11 +18,7 @@ export const newRichTextCell = (content?: string): RichTextCell => ({
 });
 
 /** A cell containing custom data, usually a formal object. */
-export type FormalCell<T> = {
-    tag: "formal";
-    id: Uuid;
-    content: T;
-};
+export type FormalCell<T> = Cell<T> & { tag: "formal" };
 
 /** Creates a formal cell with the given content. */
 export const newFormalCell = <T>(content: T): FormalCell<T> => ({
@@ -57,10 +32,7 @@ export const newFormalCell = <T>(content: T): FormalCell<T> => ({
 Stem cells are created when the user opens the "new cell" menu and are destroyed
 and replaced when a type for the new cell is selected.
  */
-export type StemCell = {
-    tag: "stem";
-    id: Uuid;
-};
+export type StemCell = Cell<unknown> & { tag: "stem" };
 
 /** Creates a new stem cell. */
 export const newStemCell = (): StemCell => ({
