@@ -2,21 +2,12 @@ import type * as Viz from "@viz-js/viz";
 import { Show, createSignal } from "solid-js";
 import { P, match } from "ts-pattern";
 
+import type { ModelJudgment } from "catlog-wasm";
 import type { ModelAnalysisProps } from "../../analysis";
-import { PanelHeader } from "../../components";
-import type { ModelJudgment } from "../../model";
+import { Foldable } from "../../components";
 import type { ModelAnalysisMeta, Theory } from "../../theory";
 import { DownloadSVGButton, GraphvizSVG, type SVGRefProp } from "../../visualization";
-import {
-    type GraphContent,
-    type GraphvizAttributes,
-    defaultEdgeAttributes,
-    defaultGraphAttributes,
-    defaultNodeAttributes,
-    graphvizEngine,
-    graphvizFontname,
-    svgCssClasses,
-} from "./graph_visualization";
+import * as GV from "./graph_visualization";
 
 import "./graph_visualization.css";
 
@@ -25,16 +16,14 @@ export function configureModelGraph(options: {
     id: string;
     name: string;
     description?: string;
-}): ModelAnalysisMeta<GraphContent> {
+}): ModelAnalysisMeta<GV.GraphConfig> {
     const { id, name, description } = options;
     return {
         id,
         name,
         description,
         component: (props) => <ModelGraph title={name} {...props} />,
-        initialContent: () => ({
-            layout: "graphviz-directed",
-        }),
+        initialContent: GV.defaultGraphConfig,
     };
 }
 
@@ -51,30 +40,31 @@ may be added in the future.
 export function ModelGraph(
     props: {
         title?: string;
-    } & ModelAnalysisProps<GraphContent>,
+    } & ModelAnalysisProps<GV.GraphConfig>,
 ) {
     const [svgRef, setSvgRef] = createSignal<SVGSVGElement>();
 
     const title = () => props.title ?? "Graph";
+    const header = () => (
+        <DownloadSVGButton
+            svg={svgRef()}
+            tooltip={`Export the ${title().toLowerCase()} as SVG`}
+            size={16}
+        />
+    );
 
     return (
         <div class="graph-visualization-analysis">
-            <PanelHeader title={title()}>
-                <DownloadSVGButton
-                    svg={svgRef()}
-                    tooltip={`Export the ${title().toLowerCase()} as SVG`}
-                    size={16}
-                />
-            </PanelHeader>
+            <Foldable title={title()} header={header()}>
+                <GV.GraphConfigForm content={props.content} changeContent={props.changeContent} />
+            </Foldable>
             <div class="graph-visualization">
                 <Show when={props.liveModel.theory()}>
                     {(theory) => (
                         <ModelGraphviz
                             model={props.liveModel.formalJudgments()}
                             theory={theory()}
-                            options={{
-                                engine: graphvizEngine(props.content.layout),
-                            }}
+                            options={GV.graphvizOptions(props.content)}
                             ref={setSvgRef}
                         />
                     )}
@@ -89,7 +79,7 @@ export function ModelGraph(
 export function ModelGraphviz(props: {
     model: ModelJudgment[];
     theory: Theory;
-    attributes?: GraphvizAttributes;
+    attributes?: GV.GraphvizAttributes;
     options?: Viz.RenderOptions;
     ref?: SVGRefProp;
 }) {
@@ -107,7 +97,7 @@ export function ModelGraphviz(props: {
 export function modelToGraphviz(
     model: ModelJudgment[],
     theory: Theory,
-    attributes?: GraphvizAttributes,
+    attributes?: GV.GraphvizAttributes,
 ): Viz.Graph {
     const nodes = new Map<string, Required<Viz.Graph>["nodes"][0]>();
     for (const judgment of model) {
@@ -119,8 +109,8 @@ export function modelToGraphviz(
                 attributes: {
                     id,
                     label: name,
-                    class: svgCssClasses(meta).join(" "),
-                    fontname: graphvizFontname(meta),
+                    class: GV.svgCssClasses(meta).join(" "),
+                    fontname: GV.graphvizFontname(meta),
                 },
             });
         }
@@ -156,8 +146,8 @@ export function modelToGraphviz(
             attributes: {
                 id: judgment.id,
                 label: judgment.name,
-                class: svgCssClasses(meta).join(" "),
-                fontname: graphvizFontname(meta),
+                class: GV.svgCssClasses(meta).join(" "),
+                fontname: GV.graphvizFontname(meta),
                 // Not recognized by Graphviz but will be passed through!
                 arrowstyle: meta?.arrowStyle ?? "default",
             },
@@ -168,8 +158,8 @@ export function modelToGraphviz(
         directed: true,
         nodes: Array.from(nodes.values()),
         edges,
-        graphAttributes: { ...defaultGraphAttributes, ...attributes?.graph },
-        nodeAttributes: { ...defaultNodeAttributes, ...attributes?.node },
-        edgeAttributes: { ...defaultEdgeAttributes, ...attributes?.edge },
+        graphAttributes: { ...GV.defaultGraphAttributes, ...attributes?.graph },
+        nodeAttributes: { ...GV.defaultNodeAttributes, ...attributes?.node },
+        edgeAttributes: { ...GV.defaultEdgeAttributes, ...attributes?.edge },
     };
 }
